[refactor] Corrigo Pastas
