--- Locacao.ts
@@ -1,6 +1,0 @@
-export class Locacao{
-    constructor(){
-
-    }
-
-}
