+++ src/models/Carro.ts
@@ -1,7 +1,7 @@
-import { Veiculo } from "./Veiculo";
+import Veiculo from "./Veiculo";
 
-export class Moto extends Veiculo {
-  public tipo: string = "A" //Moto: CNH A
+export default class Carro extends Veiculo {
+  public tipo: string = "B" //Carro: CNH B
   constructor(
     public placa: string,
     public marca: string,
