+++ src/models/Locacao.ts
@@ -1,0 +1,6 @@
+export default class Locacao{
+    constructor(){
+
+    }
+
+}
